Fix docker race between the apps and etl container
When starting the containers for the first time, docker will use the
first container that starts to fill the /etc/opensemanticsearch volume.

This volume should not be initialized like this and the apps container
has been updated to reflect that. In this commit we make sure we do not
inadvertently initialize it from the etl container either.

The new entrypoint script also waits until the apps container is finished
initializing the configuration and database.